+++ encodings/Enc_GTEIncremental.cc
@@ -155,7 +155,8 @@ void GTEIncremental::encode(Solver *S, vec<Lit> &lits, vec<uint64_t> &coeffs,
 
   // If the rhs is larger than INT32_MAX is not feasible to encode this
   // pseudo-Boolean constraint to CNF.
-  if (rhs >= INT32_MAX) {
+  // CHANGED TO INT64_MAX for now - Sukrut
+  if (rhs >= INT64_MAX) {
     printf("c Overflow in the Encoding\n");
     printf("s UNKNOWN\n");
     exit(_ERROR_);
@@ -177,8 +178,9 @@ void GTEIncremental::encode(Solver *S, vec<Lit> &lits, vec<uint64_t> &coeffs,
   for (int i = 0; i < simp_lits.size(); i++) {
     if (simp_coeffs[i] == 0)
       continue;
-
-    if (simp_coeffs[i] >= INT32_MAX) {
+	
+	// CHANGED TO INT64_MAX for now - Sukrut
+    if (simp_coeffs[i] >= INT64_MAX) {
       printf("c Overflow in the Encoding\n");
       printf("s UNKNOWN\n");
       exit(_ERROR_);
